requirements.txt
opencv-python
mediapipe
gradio
google-generativeai
transformers
torch
accelerate
scipy